requirements.txt
speechrecognition==3.10.4
PyAudio==0.2.14
google-generativeai==0.8.3
python-dotenv==1.0.1
